gles2, glsl: add gles2 support to glslparser test
Add GLSLESParserTest class which runs the gles2/gles3
version of glslparsertest. Why is this needed?
The current GLSLParserTest class doesn't know about the
gles2/gles3 varient of glslparsertest. The versioning
of glsl_version in the config section of frags and vert
shaders is conceptually incompatible for the GLSL ES
versions.

GLSLESParserTest extends the GLSLParserTest. Through
good OO, this keeps the new code to a minimum and
shares a lot of the GLSLParserTest implementation.

Why? Consider. GLSL ES version 1.00 is approx equal to
GLSL 1.20. However there are situations where
a testcase that will pass on GLSL 1.20 but will fail
under GLSL ES 1.00.

Ex: tests/glslparsertest/shaders/CorrectConstruct.vert
Embedded structure definitions are not allowed in GLSL ES
1.00 but are perfectly legal in GLSL 1.10. For both GLSL
and GLSL ES this is a good testcase.

This creates a situation where version comparisons between
verions 1.00, 1.10, 1.20, etc just isn't going to work and
expected results can be different for the same testcase.

To deal with this situation, GLSLESParserTest looks for
glsles_version in the config section. If not found, the
test is skipped. If found, this version is passed into
the glsl es version of glslparsertest.

glsles_expect_result is also added as an option entry
in the config section. If it is not present the
expect_result is the same for GLSL and GLSL ES. If
present the glsles_expect_result is passed into the
appropriate GLSL ES version of glslparsertest.